@@ -16,7 +16,7 @@
 
 package com.emmanuelkehinde.sampleapp;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.emmanuelkehinde.shutdown.Shutdown;
